@@ -7,5 +7,5 @@ interface InvestmentInterface
      * @param bool $with_commission
      * @return float
      */
-    public function getPricePaid(bool $with_commission = true);
+    public function getPricePaid(bool $with_commission = true): float;
 }
